Fix inconsistencies and broken key bindings in the Rust layer (#16217)
* [rust] Resurrect rust-backend variable

rust-backend was removed but is still meaningful (needs to be set to
the symbol lsp in order for LSP support to be enabled).

This commit restores the variable to config.el and removes boundp
checks because it will always be bound now (except in layers.el which
is loaded before config.el).

* [rust] Fully remove support for unmaintained rls backend

* [rust] Bind SPC m c . to rustic-cargo-run-rerun

This command has different behavior than cargo-process-repeat (the
previous meaning of this binding), but it is still a reasonable
binding and is the most similar command that exists in the rustic
package.

* [rust] Fix wrong command name

* [rust] Remove wrapper commands for package cargo

This package was removed in the Rust layer overhaul, so these commands
are not used anywhere.  In fact, the reloading behavior cannot easily
be implemented with rustic, because rustic runs `cargo add` and the
like in a compilation-mode buffer rather than synchronously.

As a result, we also remove references to automatically reloading the
workspace in rust-analyzer and the corresponding config variable.

* [rust] Remove unnecessary post-init stage

Layer rust owns package rustic, so there is no need to use post-init.

* [rust] Ensure rustic-mode key bindings are always available

Using :after in use-package prevents the :init forms from being
evaluated until lsp-mode and flycheck are loaded.  However, if the
user has disabled lsp-mode or flycheck, they should still be able to
use the rustic-cargo-* key bindings, so :after is not appropriate
here.

Instead, directly use with-eval-after-load where necessary for the
lsp-mode- and flycheck-specific configuration.

* [rust] Add rustic-clippy as an available flycheck checker

This just uncomments some code that was already there.  Not sure why
it was commented in the first place.

* [rust] Remove redundant progn