tell company-dabbrev to only consider same-mode buffers
With many buffers open, many of the completions offered by company-dabbrev can
be nonsense. For example, if a magit log happens to be open, then many of the
suggested completions may be hashes of commits.

set `company-dabbrev-other-buffers` to `t` which means that company-dabbrev
should only search for completions in other buffers with the same major mod