Remove outdated forks (#16300)
* [org] Remove unused local copy of ox-gfm.el

* [org] Use MELPA version of ox-gfm

The fork has fallen far behind the upstream version, and the bug it
was introduced to address was seemingly fixed by this commit:

https://github.com/larstvei/ox-gfm/commit/32dd3626b6d8efa4b1cbefc0077e737243fb18d4

(The pull request that commit claims to fix is the same one that was
cherry-picked onto the fork).

* [sql] Use ELPA version of sql-indent

sql-indent is no longer on MELPA, so Spacemacs should fetch it from
the proper GNU ELPA source instead.

* [gpu] Use MELPA version of company-glsl

Kaali's repo has been inactive for some time, and MELPA is now using
a fork thereof as the official source.