[compleseus] Use consult-theme from `theme-transient-state'
Unlike `spacemacs/helm-themes', `counsel-load-theme', and `consult-theme',
`load-theme` does not disable other themes when enabling one. It also does not
provide previews. Hence we use `consult-theme' from the `theme-transient-state`
when the compleseus layer is used; this also makes the binding consistent with
`SPC T s`.

Because consult uses the value of `this-command` to determine the `:preview-key`
property, we need to customize the wrapper `spacemacs/theme-loader' too in order
to enable previews.