[themes] add colorful and legible themes - ef-themes
The `ef-themes` are a collection of light and dark themes for GNU Emacs whose
goal is to provide colorful ("pretty") yet legible options for users who want
something with a bit more flair than the `modus-themes`.

The `ef-themes` are created by the author of `modus-themes` - the highly
accessible themes for GNU Emacs.

Screenshots: https://protesilaos.com/emacs/ef-themes-pictures