[completion] Tweak `*-spacemacs-help//layer-action-open-file`
The default layer action accessed through pressing `RET` from `SPC h l`
currently insists on creating a README.org file if none exists for the
layer which might be the case for private layers. Also, if one uses one of the
actions to directly open on of the layer files, for example keybindings.el, and
the file does not exist, only a message is displayed.

This commit changes the corresponding functions for compleseus, ivy
and helm to fall back to opening dired at the layer directory if the file does
not exist. When requesting to edit the README, it is still created if necessary.

Here is a before/after comparison of the behaviour given the
conditions (ignoring the special case of giving a prefix argument in Helm):

| is org file | edit | file exists | before                      | after                   |
|-------------+------+-------------+-----------------------------+-------------------------|
| nil         | nil  | nil         | **message**                 | **message+dired **      |
| nil         | nil  | t           | find-file                   | find-file               |
| nil         | t    | nil         | **message**                 | **message+dired**       |
| nil         | t    | t           | find-file                   | find-file               |
| t           | nil  | nil         | **spacemacs/view-org-file** | **message+dired**       |
| t           | nil  | t           | spacemacs/view-org-file     | spacemacs/view-org-file |
| t           | t    | nil         | find-file                   | find-file               |
| t           | t    | t           | find-file                   | find-file               |