[spacemacs-defaults] Clean up old linum-related code (#16462)
* Reindent some files before making substantial edits

* [spacemacs-defaults] Reindent keybindings.el

Switch from a backquoted list literal to a list function call to avoid
weird indentation alignment.

* [spacemacs-defaults] Delete unused function spacemacs/no-linum

This function would not have worked correctly anyway, as 'linum-mode
is incorrectly quoted.

* spacemacs//linum-update-window-scale-fix: Delete unused function

* [spacemacs-defaults] Refactor some duplicated code

* [spacemacs-defaults] Refactor linum to use :before-while

Also, stop calling it linum since that mode is no longer used.

* spacemacs//linum-curent-buffer-is-not-too-big: Rename and refactor

* Remove redundant display-line-numbers-mode hooks

spacemacs//linum-backward-compabitility only returns non-nil if
dotspacemacs-line-numbers is also non-nil, in which case the following
code would cause global-display-line-numbers-mode to be enabled
anyway.

* spacemacs//linum-backward-compabitility: Inline and simplify duplicated code

* spacemacs/enable-line-numbers-p: Simplify using memq

* spacemacs//linum-enabled-for-current-major-mode: Rename function

* spacemacs/derived-mode-p: Make obsolete

This is equivalent to provided-mode-derived-p, which was added in
Emacs 26.1.

* [spacemacs-defaults] Remove linum margin commands and bindings

These commands are never invoked anymore since Spacemacs does not use
linum, and therefore there is not regularly a non-zero left margin in
which to click anyway.