[org] Prevent SPC TAB from switching to `*Calendar*` (#16491)
* spacemacs/alternate-buffer: Do not select useless buffers

Similarly to how spacemacs-defaults configures next-buffer and
previous-buffer to skip useless buffers, skip them in
spacemacs/alternate-buffer (as long as the spacemacs-defaults layer is
used).

While we're at it, refactor this function to make the commonality
between the two branches more obvious.

* [org] Reindent packages.el

* [org] Prevent SPC TAB from switching to *Calendar*

This buffer is temporarily displayed whenever `org-read-date' is
called, but org-mode users do not typically use it by itself, and it
gets in the way of switching to other org buffers.