[layouts] Update list of restricted functions
This commit introduces a custom set function to make it possible to change
`spacemacs-layouts-restricted-functions` without having to restart Spacemacs.

Additionally this commit changes the default value by adding
`switch-to-next-buffer` and `switch-to-prev-buffer` (which are called by
`next-buffer` and `previous-buffer`).

Note that this only changes the behaviour of these functions in the case when they
do not find a buffer in `window-next-buffers` or `window-prev-buffers` and fall
back to cycling through the buffer list. In particular, buffers from different
layouts, which were previously (intentionally) displayed in the window could
still show up when cycling with `SPC b p` and `SPC b n` after this
change. (Though this is by default prevented with a buffer predicate, see the
option `persp-set-frame-buffer-predicate`. I don't think that's a good default,
and it is inconsistent in daemon mode, but this is a different story...)

The behaviour is only changed when cycling further than the window history, for
example using `SPC b p` and `SPC b n` in a new window will no longer switch to
buffers from different layouts.

Note that `switch-to-prev-buffer` can also get called by `quit-restore-window`,
which is another instance where, previously, buffers from different layouts could
show up unexpectedly.