[compleseus] Disable automatic preview by default
Keep it only for theme selection. For almost all other commands it was already
disabled by default. Only a seemingly random collection of a few commands had
automatic preview enabled. I think it makes more sense to leave this to users to
customize. Moving the `consult-preview-key` customization to the init block
allows overriding it in the dotfile without using
`spacemacs|use-package-add-hook` or `with-eval-after-load`.

Also make `C-M-j` and `C-M-k` preview without delay in consult commands (they
already do that for other commands).