core-themes-support.el: Run hooks after enable-theme not load-theme (#16717)
* core-themes-support.el: Run hooks after enable-theme not load-theme

The intention of spacemacs-post-theme-change-hook is to run after the
theme is changed, not just after a new theme file is loaded.
Spacemacs itself happens to always call load-theme to enable a theme,
but consult-theme (and maybe other packages) call enable-theme rather
than load-theme if a theme has already been loaded.

In any case, switching to advising enable-theme should be fine because
load-theme calls enable-theme anyway.

* [colors] Remove colors//tweak-theme-colors argument

At all callsites, this was just using the current theme, and in any
case it should not be an interactive command (which would not have
worked previously due to the wrong number of arguments).

* core-themes-support.el: Delete pointless spacemacs/post-theme-init function

This function's only purpose was to run one hook.  Its interactive
spec was wrong (did not provide any arguments) so it could not be used
interactively anyway.  Also, update the places where this function is
advised, to instead just add to that hook.