chore: remove initialization warning about pyright and lsp formatting
The warning is misleading -- the user may believe that it is not possible to
use lsp formatting when pyright is used -- this is wrong, since add-on
servers (such as ruff) may provide code formatting and be used alongside the
primary language server.

A better approach is to let the user set it up and warn only if
lsp-format-buffer fails due to lack of server capabilities.