[spacemacs-defaults] Tweak savehist configuration
1. Use an idle timer instead of a strict periodic autosave interval. Saving
large histories can cause noticable delays, and the default `history-length` set
by Spacemacs is relatively large (1000).

2. Save `kill-ring` without text properties, as they are known to cause
significant performance problems and huge history files in some cases, and
persisting them in the kill-ring usually brings little benefit.
This should fix #1300, #1369 and #9409 again.

3. Add `kmacro-ring` and `log-edit-comment-ring` as additional variables. They
are useful and usually not very large. BTW, `search-ring`, `regexp-search-ring`
and `extended-command-history` are redundant here as long as long as
`savehist-save-minibuffer-history` is non-nil (the default). I decided to leave
them there for now to avoid any breaking changes for users that customized that
option, though I guess it could be okay to change this.