[defaults] Improve `SPC w m`, persist `quit-restore` parameter
This is a follow-up to #16533, and in particular addresses the caveat mentioned
there.

The built-in tab-bar.el similarly keeps a writable window state as well as a
window configuration for each tab.

Making the `quit-restore` parameter persistent allows quitting help windows
correctly after switching tab-bar tabs, for example (without using popwin).