[modeline] Fix error when using vanilla modeline and refactor
The call to `spacemacs//restore-buffers-powerline` in
`dotspacemacs/sync-configuration-layers` would lead to an error when
using the `spacemacs-modeline` layer and the vanilla mode line.

A fix for #16346 would solve this, but in any case, this rather belongs in
spaceline package init function.

Also factor out the spaceline toggle condition into
`spacemacs//enable-spaceline-p` as it is now used in three places.