Add --no-dotspacemacs and --default-dotspacemacs options (#16958)
* [core] Add --no-dotspacemacs and --default-dotspacemacs arguments

When passed, these arguments inhibit loading the user's dotfile,
instead loading nothing, or the dotspacemacs template, respectively.

* [core] dotspacemacs/location: Error if user dotfile is not loaded

Prevent accidentally modifying the dotfile from a Spacemacs that was
started with --no-dotspacemacs or --default-dotspacemacs.  Similar to
how Emacs will refuse to modify user-init-file if started with -q.

Also use this function in preference to the dotspacemacs-filepath
variable, in most places.

* [core] Do not delete orphan packages when not loading .spacemacs

When testing the "default" Spacemacs config, it is useful to be able
to run a separate Emacs instance on the side without causing all of
your ELPA packages to be deleted.

* doc/FAQ.org: Document --default-dotspacemacs and --no-dotspacemacs