[navigation] Add cycling functionality to SPC TAB and SPC w TAB (#17161)
* [core] Refactor spacemacs/alternate-buffer

Since Emacs 27, there is the option switch-to-prev-buffer-skip which we can
leverage instead of iterating through the window-prev-buffers ourselves. The
only difference between using previous-buffer and manually switching to the
buffer is that previous-buffer respects and sets window-next-buffers. In order
to keep the same behaviour, we need to set the window's next buffers to nil
before and after invoking previous-buffer. (Note that a window's next buffers are
usually a subset of the window's previous buffers, and switching to a buffer with
any command other than previous-buffer or next-buffer will also set the window's
next buffers to nil.)

This simplifies the function a little bit, and more importantly, enables reusing
the predicate spacemacs//alternate-buffer-skip elsewhere without code
duplication.

Also, deprecate the optional window parameter. It serves no purpose, one can
just as well use with-selected-window at call sites. (And calls using the window
argument will not work with the cycling functionality from the next commit
enabled.)

* [core] Add warnings for spacemacs/alternate-buffer argument

Add a byte-compilation warning, and also add a runtime warning since
users often don't byte-compile their dotfiles.

* [navigation] Add cycling functionality to SPC TAB and SPC w TAB

Sometimes one finds oneself in the situation where one used SPC TAB to switch to
the last buffer, but it turns out that one actually intended to switch to the
second or third to last buffer. By enabling the option added in this commit, one
can consecutively press TAB in such situations to cycle through the previous
buffers. From the point of view of subsequent commands it is as though one
switched directly to the destination buffer. This uses the excellent
transient-cycles package under the
hood (https://elpa.gnu.org/packages/transient-cycles.html).

The setup needs lexical binding in spacemacs-navigation/packages.el. I have
audited the file and it does not look as if anything depends on dynamic binding.

I considered (and implemented a first version) using hydra/Spacemacs transient
states, by the way, but the implementation using the transient-cycles package is
cleaner.

* [navigation] Remove unnecessary symbol-function

Binding the variables directly to the function name is better for
debuggability, and is terser anyway.

* [navigation] Tweak buffer cycling for tab-line-mode

Use push-window-buffer-onto-prev instead of manual consing. This is simpler and
avoids duplicate entries (which do no harm and would get cleaned up lazily, but
can be visible on the tab-line).

Also, immediately update the tab-line after exiting transient cycling.

---------

Co-authored-by: Aaron L. Zeng <me@bcc32.com>