[ai] Spliting skills from personas
as feedback showed that this makes the system
much more maintainable and the personas more
powerful as now each persona knows all rules.

Before parts of the rules where part of the
various personas resulting in the coder no being
able to produce correct code on the first try
but instead had to generate code which had to be
corrected by the rules and was only right on the
second try.